+++ app/src/main/java/com/codesoom/assignment/controllers/user/UserCreateController.java
@@ -1,8 +1,6 @@
 package com.codesoom.assignment.controllers.user;
 
 import com.codesoom.assignment.application.user.UserCreator;
-import com.codesoom.assignment.application.user.UserDeleter;
-import com.codesoom.assignment.application.user.UserUpdater;
 import com.codesoom.assignment.domain.user.User;
 import com.codesoom.assignment.dto.user.UserData;
 import lombok.RequiredArgsConstructor;
@@ -14,28 +12,13 @@ import javax.validation.Valid;
 @RestController
 @RequestMapping("/users")
 @RequiredArgsConstructor
-public class UserController {
-
+public class UserCreateController {
     private final UserCreator userCreator;
-    private final UserUpdater userUpdater;
-    private final UserDeleter userDeleter;
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public User create(@RequestBody @Valid UserData userData) {
         return userCreator.createUser(userData);
-    }
-
-    @PatchMapping({"/{id}"})
-    public User update(@PathVariable Long id,
-                       @RequestBody UserData userData) {
-        return userUpdater.update(id, userData);
-    }
-
-    @DeleteMapping({"/{id}"})
-    @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void delete(@PathVariable Long id) {
-        userDeleter.delete(id);
     }
 
 }
